D3D12: Align UBO sizes to 256B.
D3D debug layer uses the descriptor size (width) to
validate CBV bounds when directly allocating UBOs.
This causes validation failure when the buffer size
is misaligned (ie. not a multiple of 256B) even
through the underlying resource heap size is always
64KB aligned.

This change always aligns the buffer size to be 256B
to avoid such validation error should sub-allocation
fail.

BUG=dawn:506

+++ src/dawn_native/BindGroupLayout.h
@@ -115,7 +115,8 @@ namespace dawn_native {
         SlabAllocator<BindGroup> MakeFrontendBindGroupAllocator(size_t size) {
             return SlabAllocator<BindGroup>(
                 size,  // bytes
-                Align(sizeof(BindGroup), GetBindingDataAlignment()) + GetBindingDataSize(),  // size
+                Align(static_cast<uint32_t>(sizeof(BindGroup)), GetBindingDataAlignment()) +
+                    GetBindingDataSize(),                                // size
                 std::max(alignof(BindGroup), GetBindingDataAlignment())  // alignment
             );
         }
